synclavix/requirements.txt
aiohttp
ccxt
python-dotenv
pandas
